recognize email header (more) correctly (than before)
Make sure that mail lines that look like header (but are not headers) are
recognized as not headers.

Initially it was only checked if the first not-empty line looks like a header, e.g.
it has a colon somwehere - however that fails with some html messages.
This adds an extra check to see if the key part of the test-header has any space:
if it has a space, it is considered to be a non-header line, and treated as part of
the message. Will see if it fails me.