parse emails with s/mime encoding
Just now I started to get emails with S/MIME signatures, which were not
handled correctly. These messages have an extra line between the header
and the actual body ("This is an S/MIME signed message"), which were not
accounted for. This change makes this line less surprising, and let's the
emails to be parsed.