handline `Content-Disposition: inline` header
In case a mailPart contains the mentioned header, it should be displayed
along with the main content. E.g mails on the Yocto mailing list
contain usually 2 `TEXT` mailparts, one is for the main message, and
one as a generic footer. The footer is sent with inline Content-
Disposition header, however the main logic handles only 1 TEXT and/or
1 HTML content, which made the footer to overwrite the main content.

This checks if there are any inline content present, and if yes, it merges
the inline content with its non-inline counterpart before persisting in db.