handle `Content-Disposition: inline` header - take 2
The previous shot at handling it was kinda working, but wasn't without
flaws: in case the encoding of the message parts were different, only
the main part was decoded correctly, and the inline part was handled
the same way, even if that was incorrect.

This change reverts most of those changes, and substitutes them with
a new attempt: keep handling the emails parts separately, and decode them
also independently from each other. When persisting it on the disk in
order to display them, in case the file being created already exists
(`index.html` or `index.txt`), don't overwrite it with the new mailpart,
but rather append it following a newline.