speed up email retrieval, remove race condition
1. This commit has some minor cleanups, like adding `std::optional` instead of
spinkling around (more) empty strings, removed some obsolete/unused functions.

2. The main thing is however a fix for a crash, that happened in some unfortunate
cases when the curl-engine was fetching emails, and the UI has switched folder.
When that happened, it was possible that when querying the data the db was in an
inconsistent state, regarding the latest email that was still being fetched.
This was solved in 2 ways:

Way1: speed up the email querying by only querying the email metadata (without
content) when it is displayed on the email list only. The content itself is loaded
in a lazy fashion, only before it would be written to disk for WebView to display.
This has increased the application performance significantly. Though it is not
noticable on desktop, now the PinePhone performance is easily on par also.

Way2: though most probably Way1 solved the issue by speed, the main DB operations are
also guarded by a mutex, just to make sure they don't start quarelling with
each other.