db schema: add update and migration feature. Update to db schema v2.
Surprise, surprise: the very first version of db-schema is not the best thing
ever. When it needs to be modified however, it would be nice if the db content
would not have to be nuked.

This change adds this: upon startup always check the current db-schema version,
and if needed, run any migration scripts.

This change also comes with the new, version 2 of the db schema: upon start it
automatically separates the `mails` table's `sender` field to `sender_name` and
`sender_email`. During migration the original `sender` field's content is also
separated, and the correct fields are populated.