display namefield only for sender - new attempt
The last attempt from yesterday worked, but it was just not good, at
least implementation-wise. Doing such parsing in the `MailModel`? Haha, no.

This commit separates the sender mail and email right after fetching the
email, and stores it properly in the db, instead of just one blob of string.