add UTF-7 decoding to display foldernames correctly
As the subject says: exotic characters in folder names are in UTF-7
encoding. This commit adds a basic UTF-7 decoder, so the folder names
are displayed correctly[*], even if they contain some exotic characters.

[*]: There must be a dedicated circle in hell for the inventors of
character encodings. Current code works for my characters in the
folder names, for the time being. Though 90-95% of this commit comes
from sensible ideas, and works as one would expect, the remaining
5-10% just happens to work for some reason.