separate readable and canonical folder names
While decoded folder names are pleasing to the eye, they are not so
pleasing for the backend, which expects the original, ugly foldername
when switching folders, and the data is fetched from the server/database.

Due to this, the folder names are stored in two forms (not in DB yet):
in the original format, as they arrived, and in decoded/readable format also.

The readable format is only used on the UI, otherwise the original format is
used for everything else.