store folder name in original and decode form also
This commit contains some extra thoughts about the previously added
folder name handling (decoded UTF-7). Instead of storing the folder list
in a bare vector in `FolderModel`, the correct names are queried from
the database one by one. I trust that this should not introduce
noticable performance degradation, considering that the size of the
`folders` table is less than 3 dozen rows.

This allows storing and passing only the folder index around instead of
juggling with readable and canonical names at the same time, in multiple models:
just remember the index, and let the DB handle the rest.