fix parsing headers with no space
The original implementation expected to have a space between the header-key
and header-value, beside the colon, so basically the following format
was expected: `header-key: header-value`.

However (notably in some spam emails) some headers don't have a space after the colon,
so they look like this: `header-key:header-value`, which made the parser to cut
off the first character of the header value. This was especially problematic with
base64 and quoted-printable encodings.

This change makes the parser not to care about the space after the colon: simply
treat the part after the colon as the header-value, and just trim it before sending
it forward.