remove focus from drawer
When the drawer is opened the first time after the email application
was displayed on the screen (either freshly started, or switched to
the email application from another application), and the drawer was opened,
wlroots recognized it as a text input receiving active focus, and it
displayed the virtual keyboard.

Not sure still why it is recognized as a text input, nevertheless, disabling
active focus-request in the qml seems to solve this problem, the keyboard
is not displayed anymore when the drawer gets focus.